Move mtls test helpers into a new package for reusability.
